Format most text files
All the changes included here are automated by using a few tools.
Some of the most important ones are mentioned below.

For python we are using black, which became very popular for being
simple (it is intentionally opinionated to prevent discussions about
style). We are also using isort for sorting imports which is useful
especially to reduce conflicts.

For most other files format, we are using prettier, which is also
somewhat opinionated and supports a large number of file formats.
This way we can format all yaml, json, md files without having to
use a separate tool for each of them.